+++ src/main/java/com/buuz135/industrial/utils/apihandlers/plant/BlockNetherWartRecollectable.java
@@ -1,6 +1,6 @@
 package com.buuz135.industrial.utils.apihandlers.plant;
 
-import com.buuz135.industrial.api.plant.IPlantRecollectable;
+import com.buuz135.industrial.api.plant.PlantRecollectable;
 import net.minecraft.block.BlockNetherWart;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.item.ItemStack;
@@ -10,7 +10,11 @@ import net.minecraft.world.World;
 
 import java.util.List;
 
-public class BlockNetherWartRecollectable implements IPlantRecollectable {
+public class BlockNetherWartRecollectable extends PlantRecollectable {
+
+    public BlockNetherWartRecollectable() {
+        super("blocknetherwart");
+    }
 
     @Override
     public boolean canBeHarvested(World world, BlockPos pos, IBlockState blockState) {